+++ Source/ACE/Entity/Container.cs
@@ -72,34 +72,39 @@ namespace ACE.Entity
             ActionChain actionChain = new ActionChain();
             actionChain.AddAction(this, () =>
             {
-                if (Inventory.ContainsKey(inventoryItem.Guid))
-                {
-                    // if item exists in the list, we are going to shift everything greater than the moving item down 1 to reflect its removal
-                    if (inventoryItem.UseBackpackSlot)
-                        Inventory.Where(i => Inventory[inventoryItem.Guid].Placement != null &&
-                                             i.Value.Placement > (uint)Inventory[inventoryItem.Guid].Placement &&
-                                             i.Value.UseBackpackSlot).ToList().ForEach(i => i.Value.Placement--);
-                    else
-                        Inventory.Where(i => Inventory[inventoryItem.Guid].Placement != null &&
-                                             i.Value.Placement > (uint)Inventory[inventoryItem.Guid].Placement &&
-                                             !i.Value.UseBackpackSlot).ToList().ForEach(i => i.Value.Placement--);
-
-                    Inventory.Remove(inventoryItem.Guid);
-                }
-                // If not going on the very end (next open slot), make a hole.
-                if (inventoryItem.UseBackpackSlot)
-                    Inventory.Where(i => i.Value.Placement >= placement &&
-                                         i.Value.UseBackpackSlot).ToList().ForEach(i => i.Value.Placement++);
-                else
-                    Inventory.Where(i => i.Value.Placement >= placement &&
-                     !i.Value.UseBackpackSlot).ToList().ForEach(i => i.Value.Placement++);
-
-                inventoryItem.Placement = placement;
-                inventoryItem.Location = null;
-                AceObject aceO = inventoryItem.SnapShotOfAceObject();
-                Inventory.Add(inventoryItem.Guid, aceO);
+                AddToInventoryEx(inventoryItem, placement);
             });
             actionChain.EnqueueChain();
+        }
+
+        public virtual void AddToInventoryEx(WorldObject inventoryItem, uint placement = 0)
+        {
+            if (Inventory.ContainsKey(inventoryItem.Guid))
+            {
+                // if item exists in the list, we are going to shift everything greater than the moving item down 1 to reflect its removal
+                if (inventoryItem.UseBackpackSlot)
+                    Inventory.Where(i => Inventory[inventoryItem.Guid].Placement != null &&
+                                         i.Value.Placement > (uint)Inventory[inventoryItem.Guid].Placement &&
+                                         i.Value.UseBackpackSlot).ToList().ForEach(i => i.Value.Placement--);
+                else
+                    Inventory.Where(i => Inventory[inventoryItem.Guid].Placement != null &&
+                                         i.Value.Placement > (uint)Inventory[inventoryItem.Guid].Placement &&
+                                         !i.Value.UseBackpackSlot).ToList().ForEach(i => i.Value.Placement--);
+
+                Inventory.Remove(inventoryItem.Guid);
+            }
+            // If not going on the very end (next open slot), make a hole.
+            if (inventoryItem.UseBackpackSlot)
+                Inventory.Where(i => i.Value.Placement >= placement &&
+                                     i.Value.UseBackpackSlot).ToList().ForEach(i => i.Value.Placement++);
+            else
+                Inventory.Where(i => i.Value.Placement >= placement &&
+                 !i.Value.UseBackpackSlot).ToList().ForEach(i => i.Value.Placement++);
+
+            inventoryItem.Placement = placement;
+            inventoryItem.Location = null;
+            AceObject aceO = inventoryItem.SnapShotOfAceObject();
+            Inventory.Add(inventoryItem.Guid, aceO);
         }
 
         public bool HasItem(ObjectGuid inventoryItemGuid, bool includeSubContainers = true)