+++ Source/ACE/Entity/Player.cs
@@ -20,7 +20,7 @@ using System.Diagnostics;
 
 namespace ACE.Entity
 {
-    public sealed class Player : Creature
+    public sealed class Player : Creature, IPlayer
     {
         private static readonly ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
 
@@ -103,7 +103,7 @@ namespace ACE.Entity
             Character.CharacterOptions2Mapping = options2;
         }
 
-        private Dictionary<Skill, CreatureSkill> Skills
+        public Dictionary<Skill, CreatureSkill> Skills
         {
             get { return AceObject.AceObjectPropertiesSkills; }
         }
@@ -1122,6 +1122,16 @@ namespace ACE.Entity
                 TrackObject(wo);
                 UpdatePlayerBurden();
             }).EnqueueChain();
+        }
+
+        public void AddNewItemToInventory(uint weenieClassId)
+        {
+            var wo = Factories.WorldObjectFactory.CreateNewWorldObject(weenieClassId);
+            wo.ContainerId = Guid.Full;
+            wo.Placement = 0;
+            AddToInventory(wo);
+            TrackObject(wo);
+            UpdatePlayerBurden();
         }
 
         /// <summary>
@@ -2413,6 +2423,16 @@ namespace ACE.Entity
             inContainerChain.EnqueueChain();
         }
 
+        /// <summary>
+        /// Context: only call when in the player action loop
+        /// </summary>
+        public void DestroyInventoryItem(WorldObject wo)
+        {
+            RemoveFromInventory(wo.Guid);
+            Session.Network.EnqueueSend(new GameMessageRemoveObject(wo));
+            Session.Network.EnqueueSend(new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.EncumbranceVal, (uint)Burden));
+        }
+
         public void HandleActionDropItem(ObjectGuid itemGuid)
         {
             ActionChain dropChain = new ActionChain();
@@ -2476,6 +2496,31 @@ namespace ACE.Entity
             });
 
             dropChain.EnqueueChain();
+        }
+        
+        public void HandleActionUseOnTarget(ObjectGuid sourceObjectId, ObjectGuid targetObjectId)
+        {
+            ActionChain chain = new ActionChain(this, () =>
+            {
+                WorldObject invSource = GetInventoryItem(sourceObjectId);
+                WorldObject invTarget = GetInventoryItem(targetObjectId);
+
+                if (invTarget != null)
+                {
+                    // inventory on inventory, we can do this now
+                    RecipeManager.UseObjectOnTarget(this, invSource, invTarget);
+                }
+                else
+                {
+                    ActionChain landblockChain = new ActionChain();
+                    CurrentLandblock.ChainOnObject(landblockChain, targetObjectId, (WorldObject theTarget) =>
+                    {
+                        RecipeManager.UseObjectOnTarget(this, invSource, theTarget);
+                    });
+                    landblockChain.EnqueueChain();
+                }
+            });
+            chain.EnqueueChain();
         }
 
         public void HandleActionUse(ObjectGuid usedItemId)
@@ -2859,6 +2904,11 @@ namespace ACE.Entity
                 ret = true;
 
             return ret;
+        }
+
+        public void SendUseDoneEvent()
+        {
+            Session.Network.EnqueueSend(new GameEventUseDone(Session));
         }
     }
 }